+++ packages/core/CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `customer` relationship to Cart model
 - Added `setCustomer` method to Cart model
 - Added `setCustomer` and `getCustomer` to StorefrontSessionManager
+- Added `latestCustomer` scope to `LunarUser` trait
 - Added pipelines to PricingManager.
 - Added `isDraft()` and `isPlaced()` methods to Order model.